pyproject.toml
[project]
name = "troop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paul-Louis Pröve", email = "mail@plpp.de" }
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "pydantic-ai>=0.1.10",
]

[project.scripts]
troop = "troop:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
